+++ src/components/FollowingList/FollowingLIst.jsx
@@ -1,44 +1,14 @@
 import FollowCard from "../FollowCard/FollowCard";
-import { postFollowShips,deleteFollowShips } from "../../apis/followship";
-import { getUserFollowings } from "../../apis/user";
-import { useEffect,useState } from "react";
+import useFollow from "../../hooks/FollowHook";
+
+const FollowingList = ({users, setUsers})=>{
+    const {handleFollow, handleUnFollow} = useFollow(users,setUsers);
 
 
-const FollowingList = ({userId})=>{
-
-    const [users, setUsers] = useState([]);
-
-    const handleFollow = async(id)=>{
-        const response = await postFollowShips(id);
-        if (response && response.status === 'success'){
-            setUsers(users.map(user=>user.id ===id?{...user, isCurrentUserFollowed:true}:user));
-        }
-    };
-    const handleUnFollow = async(id)=> {
-        const response = await deleteFollowShips(id);
-        if (response && response.status ==='success'){
-            setUsers(users.map (user=>user.id===id?{...user, isCurrentUserFollowed:false}:user));
-        }
-    };
-
-    useEffect(()=>{
-        const fetchFollowers= async ()=>{
-            const userData = await getUserFollowings(userId);
-            console.log(userData);//測試
-            if (userData){
-                setUsers(userData.map(user=>({
-                    ...user.Following,
-                    isCurrentUserFollowed:user.isCurrentUserFollowed
-                })
-                ));
-            }
-        }
-        fetchFollowers();
-    },[userId]);
 
     return (
         <>
-        {users.map(user => (
+        {users&& users.map(user => (
     <FollowCard
     key ={user.id}
     user={user}
