+++ src/components/Tab/FollowTab 暫存.jsx
@@ -7,7 +7,6 @@ import FollowingList from '../FollowingList/FollowingLIst';
 //import ReplyList from '../ReplyList/ReplyList';
 import useFollow from '../../hooks/FollowHook';
 import FollowersList from '../FollowersList/FollowersList';
-import { getUserFollowings,getUserFollowers } from "../../apis/user";
 
 
 const FollowTab = ({ userId }) => {
@@ -15,33 +14,7 @@ const FollowTab = ({ userId }) => {
     const [activeTab, setActiveTab] = useState("正在追隨");
     const location = useLocation();
     const [users, setUsers] = useState([]);
-    const [followingUsers, setFollowingUsers] = useState([]);
-    const [followerUsers, setFollowerUsers] = useState([]);
 
-
-    useEffect(() => {
-        const fetchFollowersAndFollowings = async () => {
-          const followingData = await getUserFollowings(userId);
-          const followerData = await getUserFollowers(userId);
-      
-          if (followingData) {
-            setFollowingUsers(followingData.map(user => ({
-              ...user.Following,
-              isCurrentUserFollowed: user.isCurrentUserFollowed
-            })));
-          }
-      
-          if (followerData) {
-            setFollowerUsers(followerData.map(user => ({
-              ...user.Follower,
-              isCurrentUserFollowed: user.isCurrentUserFollowed
-            })));
-          }
-        }
-      
-        fetchFollowersAndFollowings();
-      }, [userId]);
-      
     useEffect(() => {
         const currentPath = location.pathname.split('/').pop();
         switch (currentPath) {
@@ -90,8 +63,8 @@ const FollowTab = ({ userId }) => {
                   
             </div>
             <Routes>
-                <Route path="followings" element={<FollowingList userId={userId} users={followingUsers}setUsers={setFollowingUsers} />} />
-                <Route path="followers" element={<FollowersList userId={userId} users={followerUsers}setUsers={setFollowerUsers}/>} />
+                <Route path="followings" element={<FollowingList userId={userId} setUsers={setUsers} />} />
+                <Route path="followers" element={<FollowersList userId={userId} setUsers={setUsers}/>} />
                 <Route path="*" element={<FollowingList userId={userId} setUsers={setUsers}/>} /> 
             </Routes>
         </div>
